requirements.txt
# Основные библиотеки для работы с PDF
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# OCR функциональность (опционально)
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image

# Работа с данными и Excel
pandas>=2.0.0
openpyxl>=3.1.0

# Переменные окружения
python-dotenv>=1.0.0

# API клиенты
requests>=2.31.0
openai>=1.0.0

# Веб-фреймворки (выберите один)
Flask>=2.3.0
flask_cors
# ИЛИ
# fastapi>=0.100.0
# uvicorn[standard]>=0.23.0

# Дополнительно для FastAPI (если выберете его)
# pydantic>=2.0.0